Rename JS files for api client
