+++ admin-dev/themes/new-theme/scss/pages/_authorization_server.scss
@@ -8,8 +8,8 @@
   }
 }
 
-#api_access_grid_panel {
-  #api_access_grid_table {
+#api_client_grid_panel {
+  #api_client_grid_table {
     .column-headers {
       th[data-column-id="active"] {
         text-align: center;
