+++ src/PrestaShopBundle/Resources/views/Admin/Configure/AdvancedParameters/AuthorizationServer/ApiClient/form.html.twig
@@ -82,5 +82,5 @@
 
 {% block javascripts %}
   {{ parent() }}
-  <script src="{{ asset('themes/new-theme/public/api_access_form.bundle.js') }}"></script>
+  <script src="{{ asset('themes/new-theme/public/api_client_form.bundle.js') }}"></script>
 {% endblock %}
