+++ src/PrestaShopBundle/Resources/views/Admin/Configure/AdvancedParameters/AuthorizationServer/ApiClient/index.html.twig
@@ -52,5 +52,5 @@
 {% block javascripts %}
   {{ parent() }}
   <script src="{{ asset('themes/default/js/bundle/pagination.js') }}"></script>
-  <script src="{{ asset('themes/new-theme/public/api_access.bundle.js') }}"></script>
+  <script src="{{ asset('themes/new-theme/public/api_client.bundle.js') }}"></script>
 {% endblock %}
